relocate ProxyConfiguration in app structure
