+++ app/src/main/java/com/keylesspalace/tusky/components/preference/ProxyPreferencesFragment.kt
@@ -19,8 +19,8 @@ import android.os.Bundle
 import androidx.preference.Preference
 import androidx.preference.PreferenceFragmentCompat
 import com.keylesspalace.tusky.R
-import com.keylesspalace.tusky.entity.ProxyConfiguration
 import com.keylesspalace.tusky.settings.PrefKeys
+import com.keylesspalace.tusky.settings.ProxyConfiguration
 import com.keylesspalace.tusky.settings.makePreferenceScreen
 import com.keylesspalace.tusky.settings.switchPreference
 import com.keylesspalace.tusky.settings.validatedEditTextPreference
