+++ app/src/main/java/com/keylesspalace/tusky/settings/ProxyConfiguration.kt
@@ -1,4 +1,4 @@
-package com.keylesspalace.tusky.entity
+package com.keylesspalace.tusky.settings
 
 import java.net.IDN
